requirements.txt
asgiref==3.8.1
cloudinary==1.40.0
coverage==7.5.3
dj-database-url==2.1.0
dj-rest-auth==6.0.0
Django==5.0.6
django-cloudinary-storage==0.3.0
django-extensions==3.2.3
djangorestframework==3.15.1
gunicorn==22.0.0
iniconfig==2.0.0
pillow==10.3.0
pluggy==1.5.0
psycopg2==2.9.9
pytest==8.2.1
pytz==2024.1
sqlparse==0.5.0
